fix: change some nulls to not nulls on migrations

+++ migrations/2024-04-20-144438_create_payment/up.sql
@@ -2,7 +2,7 @@
 CREATE TABLE payment (
     id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
     amount DECIMAL(10, 2) NOT NULL,
-    month DECIMAL(10, 2) NOT NULL,
+    month INTEGER NOT NULL,
     year INTEGER NOT NULL,
     employee_id INTEGER NOT NULL,
     FOREIGN KEY (employee_id) REFERENCES employee(id)
